refactor other components
