+++ src/components/footers/footer.jsx
@@ -1,6 +1,6 @@
 import { EmailField } from '../emails'
 import { LinkButton } from '../buttons'
-import SocialMedia from '../socialMedia'
+import { SocialMedia } from '../others'
 
 function Footer() {
   return (
