+++ src/components/navbars/navbar.jsx
@@ -6,7 +6,7 @@ import { Link, useNavigate } from 'react-router-dom'
 import { useDarkMode } from '../../hooks/useDarkMode'
 import { useNavMode } from '../../hooks/useNavMode'
 import useScrollPosition from '../../hooks/useScrollPosition'
-import BannerCarousel from '../bannerCarousel'
+import BannerCarousel from '../others/bannerCarousel'
 import { ButtonIcon, LinkButton } from '../buttons'
 import { selectUserUid } from '../../redux/features/userSlice'
 import { signOut } from '../../services/auth/index'
