+++ src/components/others/index.js
@@ -1,0 +1,17 @@
+import BannerCarousel from './bannerCarousel'
+import Counter from './counter'
+import InfoItem from './infoItem'
+import InfoItemList from './infoItemList'
+import Overview from './overview'
+import SocialMedia from './socialMedia'
+import Stepper from './stepper'
+
+export {
+  BannerCarousel,
+  Counter,
+  InfoItem,
+  InfoItemList,
+  Overview,
+  SocialMedia,
+  Stepper
+}
